requirements.txt
fastapi
uvicorn
mysql-connector-python
python-multipart
bcrypt
python-jose
python-dotenv
Babel
stripe